% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_efficacy.R
\name{add_efficacy}
\alias{add_efficacy}
\title{Add gRNA efficacy to each gRNA}
\usage{
add_efficacy(
  gRNAs,
  rule_set = "CRISPRscan",
  BSgenomeName,
  output_dir,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{gRNAs}{DNAStringSet object containing a set of gRNAs. Please note the
sequences must contain PAM appended after gRNAs, e.g.,
ATCGAAATTCGAGCCAATCCCGG where ATCGAAATTCGAGCCAATCC is the gRNA and CGG is
the PAM}

\item{rule_set}{Specify a rule set scoring system for calculating gRNA
efficacy. Please note that Root_RuleSet2_2016 requires the following python
packages with specified verion and python 2.7.  1. scikit-learn 0.16.1 2.
pickle 3. pandas 4. numpy 5. scipy}

\item{BSgenomeName}{BSgenome object. Please refer to available.genomes in
BSgenome package. For example,
\itemize{
\item{BSgenome.Hsapiens.UCSC.hg19} - for hg19,
\item{BSgenome.Mmusculus.UCSC.mm10} - for mm10
\item{BSgenome.Celegans.UCSC.ce6} - for ce6
\item{BSgenome.Rnorvegicus.UCSC.rn5} - for rn5
\item{BSgenome.Drerio.UCSC.danRer7} - for Zv9
\item{BSgenome.Dmelanogaster.UCSC.dm3} - for dm3
}}

\item{output_dir}{the directory where the off target analysis and reports
will be written to}

\item{overwrite}{overwrite the existing files in the output directory or}

\item{...}{Please see additional parameters in offTargetAnalysis of
CRISPRseek pacakage
not, default FALSE}
}
\value{
gRNA efficacy
}
\description{
Add gRNA efficacy to each gRNA
}
\examples{
 library(CRISPRseek)
	library("BSgenome.Hsapiens.UCSC.hg19")
	output_dir <- getwd()

}
\seealso{
CRISPRseek
}
\author{
Lihua Julie Zhu
}
