% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{aggregateGuides}
\alias{aggregateGuides}
\title{Aggregate counts or log-ratios at the gene-level}
\usage{
aggregateGuides(
  se,
  guess.log.transform = FALSE,
  aggregate.field = "group",
  fun = c("median", "mean"),
  aggregate.ntcs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing guide-level data
(counts or log-ratios).}

\item{guess.log.transform}{Should appropriate log-transformations prior
to aggregation be guessed from the data? FALSE by default.}

\item{aggregate.field}{Column name in rowData specifying the
grouping levels.}

\item{fun}{Either "median" (default) or "mean". The summary statistic used 
to calculate gene-level data.}

\item{aggregate.ntcs}{Should non-targeting controls (NTCs) be aggregated?
FALSE dy default.}

\item{verbose}{Should messaged be printed to console? TRUE by default.}
}
\value{
A SummarizedExperiment object with gene-level data. 
    The metadata for the SummarizedExperiment will also include
    "aggregate" in the \code{transform} list.
}
\description{
Computes a summary statistic of counts or log-ratios across
all guides for all genes and returns a SummarizedExperiment object with 
gene-level aggregated values.
}
