% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CRISPR_screening_data.R
\name{analyze_CRISPR_screen_data}
\alias{analyze_CRISPR_screen_data}
\title{A work flow function for merging multiple groups, statistical analysis,
and filtering results based on p-value/adjusted p-values and log odds ratio.}
\usage{
analyze_CRISPR_screen_data(
  gRNA_count_files,
  sample_names = c("A", "B", "C"),
  sep = " ",
  header = FALSE,
  comparisons_group1 = c("B", "C"),
  comparisons_group2 = c("A", "A"),
  multiAdjMethod,
  min_total_count = 6,
  maxP = 0.05,
  gene_col = 2,
  min_odds_ratio = 1,
  max_odds_ratio,
  ngRNAs_per_gene = 4,
  split = ":",
  column_to_split = "gene",
  new_column_names = c("Symbol", "ID", "gRNA"),
  drug_target,
  output_file = "allgRNAs_stats_results.xlsx",
  ...
)
}
\arguments{
\item{gRNA_count_files}{A vector of file names. Each file contains a list of
gRNAs with count information.}

\item{sample_names}{Sample names corresponding to each file in
gRNA_count_files with the same order}

\item{sep}{delimiter in the gRNA_count_files. Default to space}

\item{header}{TRUE or FALSE indicating whether the files specified in
gRNA_count_files contain header or not. Default to FALSE}

\item{comparisons_group1}{The denominator for calculating Log odds ratio}

\item{comparisons_group2}{The numerator for calculating log odds ratio}

\item{multiAdjMethod}{A vector of character strings containing the names of
the multiple testing procedures for which adjusted p-values are to be
computed. This vector should include any of the following:
"Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
and "TSBH". Please type ?multtest::mt.rawp2adjp for details}

\item{min_total_count}{gRNAs with total count summed over all groups less
than min_total_count are excluded from further statistical analysis.}

\item{maxP}{maximum p-value cutoff. Please note that if multiAdjMethod is set,
this is the maximum adjusted p-value}

\item{gene_col}{For calculating the number of significant gRNAs per gene.
Default to 2, i.e., the second column contains the gene symbol.}

\item{min_odds_ratio}{gRNAs with odds_ratio >= min_odds_ratio will be kept in
the filtered results, default to 1}

\item{max_odds_ratio}{gRNAs with odds_ratio <= max_odds_ratio will be kept in
the filtered results}

\item{ngRNAs_per_gene}{The designed number of gRNAs per gene in the
screening library}

\item{split}{Used to extract gene symbol, gene_id, and gRNA from c
olumn_to_split of gRNA_count_files. Set split to ":" for extracting gene symbol,
gene ID, and gRNA from : separated string, e.g.,
TMEM192:NM_001100389.1:TGGGTCGTCTTCAATACTCT.}

\item{column_to_split}{The name of the column that contains the
annotation. Default to "gene".}

\item{new_column_names}{Default to c("Symbol", "ID", "gRNA") if
the annotation in column_to_split is in the form of Symbol:ID:gRNA such as
MEM192:NM_001100389.1:TGGGTCGTCTTCAATACTCT, where split is ":"}

\item{drug_target}{a data frame with drug target information. The one
included in this package is from drug bank 2018
(https://go.drugbank.com/targets). Users can input their #' own drug
target data frame. The latest release can be found at
https://go.drugbank.com/releases/latest (access approval required)}

\item{output_file}{The output file name}

\item{...}{Set additional parameters for the funciton is_druggable}
}
\value{
A list of two objects with the first one being a data frame
containing all gRNAs >=  min_total_count, and the second one being a list
containing four data frames: genes with at least one gRNA, two, three,
and four gRNAs being significant respectively. Significance is determined
using the filtering criteria set by users. The filtering criteria
includes multiAdjMethod, min_total_count, maxP, min_odds_ratio,
and max_odds_ratio. Default settings will be used if users do not modify
these parameters.
}
\description{
A work flow function for merging multiple groups, statistical analysis,
and filtering results based on p-value/adjusted p-values and log odds ratio.
}
\examples{
gRNA_count_files<- c(system.file("extdata", c("A.txt", "B.txt"),
            package = "CRISPRscreen"))
drug_target <- readRDS(system.file("extdata", "drug_targets_2018.RDS",
            package = "CRISPRscreen"))
results <- analyze_CRISPR_screen_data(gRNA_count_files,
            sample_names = c("A", "B"),
            comparisons_group1 = "B",
            comparisons_group2 = "A",
            multiAdjMethod = "BH",
            min_total_count = 1, drug_target = drug_target,
            output_file = "allgRNAs_stats_results.xlsx")

}
\author{
Lihua Julie Zhu
}
