% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeEnrichment.R
\name{ComputeEnrichment}
\alias{ComputeEnrichment}
\title{ComputeEnrichment}
\usage{
ComputeEnrichment(data.merged = P5.P6.readsnumber,
  FDRmethod = "qvalue")
}
\arguments{
\item{data.merged}{The data prepared by "PrepareData"}

\item{FDRmethod}{The method used to adjust p value. "qvalue" or "bonferroni" or "none"}
}
\value{
Return the data.merged that include p/fdr and fold change in the last column
}
\description{
This function allows you to compute enrichment fold change and p value(FDR) or each
}
\examples{
ComputeEnrichment(data.merged=P5.P6.readsnumber,FDRmethod="qvalue")
}
