% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{drawDensities}
\alias{drawDensities}
\title{Generate plot of barcode distributions on a log2 scale (static plot)}
\usage{
drawDensities(
  object,
  log.scale = TRUE,
  sort = TRUE,
  col = NULL,
  col.group.field = "Group",
  lty = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "Log2 sgRNA Counts",
  ylab = "Density",
  main = "",
  legend.where = "topleft",
  assay_index = 1,
  ...
)
}
\arguments{
\item{object}{Either a numeric matric or a SummarizedExperiment}

\item{log.scale}{Should the data be log-transformed? TRUE by default.}

\item{sort}{Should samples be drawn by color order?}

\item{col}{Optional character vector specifying sample color. Must be}

\item{col.group.field}{Optional string indicating which column of colData(object) should be
used to color densities.}

\item{lty}{Optional character vector specifying sample line type. Must be 
of the same length as the number of columns in the provided object.}

\item{xlim}{Usual xlim argument in base plot}

\item{ylim}{Usual ylim argument in base plot}

\item{xlab}{Usual xlab argument in base plot}

\item{ylab}{Usual ylab argument in base plot}

\item{main}{Usual main argument in base plot}

\item{legend.where}{Where should be the legend?}

\item{assay_index}{Integer specifying which assay should be used. 1 by default.}

\item{...}{Other arguments to passe to base plot function}
}
\value{
Returns nothing. A plot is produced as a side effect.
}
\description{
Generate plot of barcode distributions on a log2 scale (static plot).
}
\examples{
\dontrun{
 data("seExample")
 drawDensities(seExample)
}
}
