% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractBarcodes}
\alias{extractBarcodes}
\title{Quick function to extract barcodes from a list of reads}
\usage{
extractBarcodes(
  reads,
  flank5 = "AGTTCG",
  flank3 = "TTCGGACTGT",
  leftOnly = FALSE,
  barcode_len = 19,
  ignore_barcode_len = FALSE
)
}
\arguments{
\item{reads}{Character vector of sequencing reads.}

\item{flank5}{String containing the constant sequence on the 5'
flank of the barcode region.}

\item{flank3}{String containing the constant sequence on the 3'
flank of the barcode region.}

\item{leftOnly}{Should only flank5 be considered? 
FALSE by default.}

\item{barcode_len}{Length of the barcodes to be retrieved.}

\item{ignore_barcode_len}{Should any barcode length be returned?
FALSE by default.}
}
\value{
Character vector of barcodes.
}
\description{
Quick function to extract barcodes from a list of reads for testing 
    purposes.
}
\examples{
\dontrun{
    reads <- extractReads(ngs="ngs3565")
    barcodes <- extractBarcodes(reads, flank5="TACCG", flank3="")
}
}
\author{
Jean-Philippe Fortin
}
