% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractReads}
\alias{extractReads}
\title{Quick function to extract a subset of reads from fastq files}
\usage{
extractReads(file, n_reads = 1000, sample_index = 1, read_mate = c("R1", "R2"))
}
\arguments{
\item{file}{String specifying path to a FASTQ file from which reads will
be read from.}

\item{n_reads}{Integer specifying number of reads that should be returned.
1000 by default.}

\item{sample_index}{Integer specifying which sample should be considered
from the list of files.}

\item{read_mate}{String specifying if reads from R1 or R2 files should
be returned. Must be either "R1" or "R2". R1" by default.}
}
\value{
Character vector of sequencing reads.
}
\description{
Quick function to extract a subset of reads from fastq files.
}
\author{
Jean-Philippe Fortin
}
