% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gRNAs.R
\name{filter_gRNAs}
\alias{filter_gRNAs}
\title{filter gRNAs}
\usage{
filter_gRNAs(
  gRNAs,
  comparison,
  multi_adj_method = "BH",
  maxP = 0.05,
  min_odds_ratio = 1,
  max_odds_ratio,
  ngRNAs_per_gene = 4,
  gene_col = 2,
  output_file = "enrichedInC.xlsx"
)
}
\arguments{
\item{gRNAs}{Results from DE_gRNAs}

\item{comparison}{The comparison name. For example, use "B vs A" for comparing
sample B and A}

\item{multi_adj_method}{A vector of character strings containing the names of
the multiple testing procedures for which adjusted p-values are to be
computed. This vector should include any of the following:
"Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
and "TSBH". Please type ?multtest::mt.rawp2adjp for details}

\item{maxP}{maximum p-value cutoff. Please note that if multi_adj_method is set,
this is the maximum adjusted p-value}

\item{min_odds_ratio}{gRNAs with odds_ratio >= min_odds_ratio will be kept in
the filtered results, default to 1}

\item{max_odds_ratio}{gRNAs with odds_ratio <= max_odds_ratio will be kept in
the filtered results}

\item{ngRNAs_per_gene}{The designed number of gRNAs per gene in the
screening library}

\item{gene_col}{For calculating the number of significant gRNAs per gene.
Default to 2, i.e., the second column contains the gene symbol.}

\item{output_file}{The output file name
@return A list of four data frames: genes with at least one gRNA, two, three,
and four gRNAs being significant respectively. Significance is determined
using the filtering criteria set by users. The filtering criteria
includes multi_adj_method, maxP, min_odds_ratio,
and max_odds_ratio. Default settings will be used if users do not modify
these parameters.}
}
\description{
filter gRNAs
}
\author{
Lihua Julie Zhu
}
