% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plot.R
\name{frequency_plot}
\alias{frequency_plot}
\title{Create the distribution of log odds ratio and summarized FDR for given genes
with or without pathway annotation}
\usage{
frequency_plot(
  x,
  genes_to_highlight,
  category_anno,
  anno_category_col = 1,
  anno_gene_col = 2,
  highlight_color_up = "red",
  highlight_color_down = "blue",
  gene_col = 2,
  alpha = 0.5,
  odds_ratio_col = 11,
  FDR_col = 12,
  bins = 100,
  inverse_odds_ratio = FALSE,
  plot_facet = c("by_gene", "by_category", "none"),
  gene_label_size = 12,
  dashline_color = "purple",
  FDR_label_size = 4,
  FDR_vjust = "inward",
  cutoff = 0,
  include_FDR = TRUE,
  gene_list_ordered = TRUE
)
}
\arguments{
\item{x}{a data frame such as the output from analyze_CRISPR_screen_data
with minimum of three columns: Gene Identifier, gene symbol,
odds_ratio, and adj_pvalue/FDR}

\item{genes_to_highlight}{the list of genes that will be highlighted in
the figure. When plot_facet is set to by_gene or none, genes_to_highlight
needs to be specified as a character vector.}

\item{category_anno}{If plot_facet is set to by_category, then
category_anno needs to be specified as a data frame with category specified in
anno_category_col and gene symbol specified in anno_gene_col.}

\item{anno_category_col}{If plot_facet is set to by_category, then
category_anno needs to be specified with category specified in
anno_category_col and gene symbol specified in anno_gene_col.}

\item{anno_gene_col}{If plot_facet is set to by_category, then
category_anno needs to be specified with category specified in
anno_category_col and gene symbol specified in anno_gene_col.}

\item{highlight_color_up}{the color used for highlighting the
genes specified by genes_to_highlight and up}

\item{highlight_color_down}{the color used for highlighting the
genes specified by genes_to_highlight and down}

\item{gene_col}{the column containing the gene symbol}

\item{alpha}{for specifying the percentage of data to plot in histogram,
default to 0.5 (50 percent)}

\item{odds_ratio_col}{the column where odds_ratios are stored}

\item{FDR_col}{the column where adj_pvalues or FDRs are stored}

\item{bins}{the number of bins used for histogram generation using
geom_histogram, default to 100}

\item{inverse_odds_ratio}{indicates whether to reverse the odds_ratio}

\item{plot_facet}{either by by_gene, by_category, or none for facet plotting
If plot_facet is set to by_category, then category_anno needs to be
specified with category specified in anno_category_col and gene symbol
specified in anno_gene_col.}

\item{gene_label_size}{the size of gene label}

\item{dashline_color}{the color for drawing the dashed line for log odds
ratio of 0}

\item{FDR_label_size}{the size of FDR label}

\item{FDR_vjust}{specifies justification orthogonal to the direction of the
FDR value, default to 0.5 (middle)}

\item{cutoff}{specify where to draw to vertical dash line to indicate the
cutoff. Default to log odds ratio of 0.}

\item{include_FDR}{Indicates whether to include FDR on the right side of
the plot for each gene. Default to TRUE.}

\item{gene_list_ordered}{Indicates whether the genes_to_highlight is ordered.
Default to TRUE, i.e., genes_to_highlight will be plotted from top to bottom.}
}
\value{
a list of ggplot object of size 2
}
\description{
Create the distribution of log odds ratio and summarized FDR for given genes
with or without pathway annotation
}
\examples{

# add category on the right side of the bottom figure by setting
# plot_facet to by_category
# load the example category data frame

category_anno <- readRDS(system.file("extdata", "example_category_ann.RDS",
            package = "CRISPRscreen"))

x <- readRDS(system.file("extdata", "example_data.RDS",
            package = "CRISPRscreen"))

p <- frequency_plot(x = x, category_anno = category_anno[1:6,],
                   genes_to_highlight = category_anno[1:6,2],
                   plot_facet = "by_category")
p[[1]]/p[[2]]

# Set plot_facet to by_gene
genes_to_highlight <- category_anno[1:6, 2]
p <- frequency_plot(x = x, genes_to_highlight = genes_to_highlight,
                   plot_facet = "by_gene")
 p[[1]]/p[[2]]
}
\author{
Lihua Julie Zhu
}
