% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenehitsCaller.R
\name{GenehitsCaller}
\alias{GenehitsCaller}
\title{GenehitsCaller}
\usage{
GenehitsCaller(enrichment.data = P5.P6.enrichment, Tier1Cut = 0.001,
  Tier2Cut = 0.01, Tier1Nmin = 1, TotalNmin = 2,
  conflict = "Onlytier1")
}
\arguments{
\item{Tier1Cut}{The pvalue/FDR cutoff below which an sgRNA is considered tier1 sgRNA(very significant)}

\item{Tier2Cut}{The pvalue/FDR cutoff between which to Tier1Cut an sgRNA is considered tier2 sgRNA(Somewhat significant)}

\item{Tier1Nmin}{Minimum number of supporting Tier1 sgRNA required to identify a gene hit (1,2,3,4,5,6)}

\item{TotalNmin}{Minimum number of supporting Tier1+Tier2 sgRNA required to identify a gene hit (1,2,3,4,5,6)}

\item{conflict}{When getting rid of genes with conficting significant sgRNA. i,e. sgRNA show opposite direction. if "Onlytier1" then only consider confliction in Tier1 if "Alltiers" then consider both tier1 and tier2}
}
\value{
this will return a list of two dataframe including gene hits that show the number of supporting sgRNA; and tier1+tier2 sgRNA data of all hits genes
}
\description{
This function allows you to collapse sgRNA into gene in order to identify gene hits.
}
\examples{
sgRNAdistribute(enrichment.data=P5.P6.enrichment)
}
