% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEssentialityROC.R
\name{getEssentialityROC}
\alias{getEssentialityROC}
\title{Get ROC curve for classifying essential genes}
\usage{
getEssentialityROC(
  x,
  genes,
  rev = TRUE,
  essentialGenes,
  nonEssentialGenes,
  pseudo = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{genes}{A character vector specifying genes associated with x}

\item{rev}{Should negative values represent signal? TRUE by default.}

\item{essentialGenes}{Character vector of essential gene symbols.}

\item{nonEssentialGenes}{Character vector of non-essential gene symbols.}

\item{pseudo}{Should pseudo ROC curves be computed instead?
FALSE by default.}
}
\value{
A data.frame with 2 columns:
    \code{specificity} and \code{sensitivity}.
}
\description{
Get receiver operating characteristic (ROC) curve for 
classifying essential genes.
}
