% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneLevelSummary.R
\name{getGeneLevelSummary}
\alias{getGeneLevelSummary}
\title{Generate Gene-Level Statistical Summary for Each Gene}
\usage{
getGeneLevelSummary(
  x,
  comparisons_group1,
  comparisons_group2,
  combine_odds_ratio_method = c("median", "mean", "max", "min"),
  combine_p_value_method = c("min", "median", "mean", "max"),
  odds_ratio_threshold = 2,
  min_gRNAs_ge_odds_ratio_threshold = 3,
  multi_adj_method = "BH"
)
}
\arguments{
\item{x}{the first list item (named as stats_results) in the
analysis results from analyze_CRISPR_screening_data}

\item{comparisons_group1}{The numerator for calculating Log odds ratio}

\item{comparisons_group2}{The denominator for calculating log odds ratio}

\item{combine_odds_ratio_method}{The method to summarize gRNA-level odds ratio
to gene-level odds ratio. Supported methods are "median", "mean", "max", and "min"}

\item{combine_p_value_method}{The method to summarize gRNA-level p.value
to gene-level p-avalue. Supported methods are Supported methods are
"min", "median", "mean", and "max"}

\item{odds_ratio_threshold}{odds ratio threshold for selecting the gRNAs.
It is used to compute n.gRNA.ge.odds_ratio_threshold for each gene
which equals the number of gRNAs with odds.ratio >= odds_ratio_threshold,
default to 2}

\item{min_gRNAs_ge_odds_ratio_threshold}{For computing
diff.expressed.adjp0.05 and
diff.expressed.adjp0.01. Default 3, meaning that
at least 3 gRNAs with odds ratio >= odds_ratio_threshold and adjusted
p.value < 0.05 and < 0.01 respectively.}

\item{multi_adj_method}{A vector of character strings containing the names of
the multiple testing procedures for which adjusted p-values are to be
computed. This vector should include any of the following:
"Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
and "TSBH". Please type ?multtest::mt.rawp2adjp for details}
}
\value{
A data frame with gene-level summary including Symbol,
p.value.gene, odds.ratio.gene, n.gRNAs.ge.odds_ratio_threshold,
diff.expressed.adjp0.05,
diff.expressed.adjp0.01, and log2.odds.ratio. It can be used
to generate a plot with n.gRNAs.ge.odds_ratio_threshold as Y-axis
and log.odds.ratio as the x-axis
}
\description{
Generate Gene-Level Statistical Summary for Each Gene
}
\examples{
gRNA_count_files<- c(system.file("extdata", c("A.txt", "B.txt"),
            package = "CRISPRscreen"))
drug_target <- readRDS(system.file("extdata", "drug_targets_2018.RDS",
            package = "CRISPRscreen"))
results <- analyze_CRISPR_screen_data(gRNA_count_files,
            sample_names = c("A", "B"),
            comparisons_group1 = "B",
            comparisons_group2 = "A",
            multi_adj_method = "BH",
            min_total_count = 1, drug_target = drug_target,
            output_file = "allgRNAs_stats_results.xlsx")

geneResults <- getGeneLevelSummary(results$stats_results,
     comparisons_group1 = "B",
     comparisons_group2 = "A",
     combine_odds_ratio_method = "median",
     combine_p_value_method= "min")
 

}
