% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGO.R
\name{getGO}
\alias{getGO}
\title{Get Gene Ontology Annotation for the input gene list}
\usage{
getGO(all.genes, orgAnn, output_file, ID_type = "gene_symbol")
}
\arguments{
\item{all.genes}{A vector of gene IDs}

\item{orgAnn}{orgAnn    Organism annotation package such as org.Hs.eg.db for
human and org.Mm.eg.db for mouse, org.Dm.eg.db for fly, org.Rn.eg.db for
rat, org.Sc.eg.db for yeast, and org.Dr.eg.db for zebrafish}

\item{output_file}{output file path}

\item{ID_type}{Gene ID type}
}
\description{
Get Gene Ontology Annotation for the input gene list
}
\author{
Lihua Julie Zhu
}
