% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{getLogRatios}
\alias{getLogRatios}
\title{Get log2 fold changes of samples with respect to a reference group}
\usage{
getLogRatios(
  se,
  condition.field = NULL,
  reference.level = NULL,
  replicate.field = NULL,
  use.normalized = TRUE
)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{condition.field}{String specifying the column of \code{colData(se)}
containing the sample group information.}

\item{reference.level}{String specifying the reference level of the column
named by \code{condition.field}.}

\item{replicate.field}{String specifying the column of \code{colData(se)}
containing the replicate sample information.}

\item{use.normalized}{Should normalized counts be used? TRUE by default.}
}
\value{
A SummarizedExperiment object with log2 fold change data with
respect to the reference. By definition, reference samples will have values
of \code{0} for all guides. The metadata for the SummarizedExperiment will
also have "log_ratio" in the \code{transform} list, which is used by other
functions to determine whether the input SummarizedExperiment has (or does
not have) the necessary data transformations.
}
\description{
Computes the log2 fold changes of guides for all samples with
respect to the reference timepoint or condition and returns a
SummarizedExperiment with those values.
}
