% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCs.R
\name{getPCs}
\alias{getPCs}
\title{Get principal components.}
\usage{
getPCs(se, condition.field, reference.level, replicate.field, how.many = 5000)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{condition.field}{String specifying the name of the column in colData(se)
that specifies the experimental groups.}

\item{reference.level}{String specifying the name of
the reference group.}

\item{replicate.field}{String specifying the name of the replicate
column in \code{colData(se)}.}

\item{how.many}{Numeric value indicating how many PCs to return.}
}
\description{
Get PCs from a Summarized Experiment object.
}
