% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_druggable.R
\name{is_druggable}
\alias{is_druggable}
\title{Annotate genes with drug information}
\usage{
is_druggable(gRNAs, drug_target, match_col_name = c("Symbol", "Gene Name"))
}
\arguments{
\item{gRNAs}{a data frame with minimum one column named as the first
element in match_col_name such as the output from
analyze_CRISPR_screen_data}

\item{drug_target}{a data frame with drug target information. The one
included in this package is from drug bank 2018
(https://go.drugbank.com/targets). Users can input their #' own drug
target data frame. The latest release can be found at
https://go.drugbank.com/releases/latest (access approval required)}

\item{match_col_name}{Only Gene and Symbol supported for the first element of
match_col_name. The first element must be a valid column name in gRNAs data
frame and the second element must be a valid column name in drug_target
data frame}
}
\value{
an annotated data frame with the last column is_druggable added and the
rest of the columns are from gRNAs
}
\description{
Annotate genes with drug information
}
\examples{

x <- readRDS(system.file("extdata", "example_data.RDS",
            package = "CRISPRscreen"))
drug_target <- readRDS(system.file("extdata", "drug_targets_2018.RDS",
            package = "CRISPRscreen"))
x <- is_druggable(x[1:10,], drug_target,
         match_col_name = c("Symbol", "Gene Name"))

}
\author{
Lihua Julie Zhu
}
