% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gRNA_files.R
\name{merge_gRNA_files}
\alias{merge_gRNA_files}
\title{Merge multiple gRNA count files from different sample groups}
\usage{
merge_gRNA_files(gRNA_count_files, sample_names, sep = " ", header = FALSE)
}
\arguments{
\item{gRNA_count_files}{A vector of file names. Each file contains a list of
gRNAs with count information.}

\item{sample_names}{Sample names corresponding to each file in
gRNA_count_files with the same order}

\item{sep}{delimiter in the gRNA_count_files. Default to space}

\item{header}{TRUE or FALSE indicating whether the files specified in
gRNA_count_files contain header or not. Default to FALSE}
}
\description{
Merge multiple gRNA count files from different sample groups
}
\author{
Lihua Julie Zhu
}
