% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeGrowthRates}
\alias{normalizeGrowthRates}
\title{Normalization for differental growth in dropout screens for focused libraries}
\usage{
normalizeGrowthRates(
  se,
  reference.field = NULL,
  reference.level = NULL,
  replicate.field = NULL,
  stratum.field = NULL,
  group.field = NULL,
  group.reference = NULL,
  gene.field = NULL,
  fun = c("median", "mean"),
  genes
)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{reference.field}{String specifying the column of \code{colData(se)} containing the type of each sample (i.e., reference or not).}

\item{reference.level}{Character vector specifying the reference levels of the column named by \code{reference.field}.}

\item{replicate.field}{String specifying the column of \code{colData(se)} containing the replicate information.}

\item{stratum.field}{String specifying the column of \code{colData(se)} containing the variable for stratification}

\item{group.field}{String specifying the column of \code{colData(se)} containing the group variable for normalization}

\item{group.reference}{Character vector specifying the group to normalize to from the column named by \code{group.field}.}

\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}

\item{fun}{median or mean: function to calculate scaling factors}

\item{genes}{gene names specifying which genes should be used for estimating the scaling factor.}
}
\value{
A SummarizedExperiment with counts normalized for differential growth
}
\description{
Normalization for differential growth in dropout screens using simple
scaling using a set of essential genes
}
