% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeNeg}
\alias{normalizeNeg}
\title{Normalization with non-essential genes}
\usage{
normalizeNeg(se, gene.col = "group", nonEssentialGenes)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{gene.col}{Column in rowData(se) containg gene symbol information.}

\item{nonEssentialGenes}{Character vector of non-essential gene symbols}
}
\value{
A SummarizedExperiment with counts normalized using non-essential genes.
}
\description{
Normalization with non-essential genes.
}
