% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeNtc}
\alias{normalizeNtc}
\title{Normalization with non-targeting controls (NTCs)}
\usage{
normalizeNtc(se, gene.col = "group")
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{gene.col}{Column in rowData(se) containg NTC information.}
}
\value{
A SummarizedExperiment with counts normalized using NTCs.
}
\description{
Normalization with non-targeting controls (NTCs).
}
\examples{
\dontrun{
se <- normalizeNtc(seExample)
}
}
