% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeTotal}
\alias{normalizeTotal}
\title{Normalization for library size}
\usage{
normalizeTotal(se, fun = c("median", "mean"))
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{fun}{Median or mean?}
}
\value{
A SummarizedExperiment with counts normalized
}
\description{
Normalization for library size
}
\examples{
\dontrun{
se <- normalizeTotal(seExample)
}
}
