% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{PrepareData}
\usage{
PrepareData(readsnumber.1 = P5.readsnumber,
  readsnumber.2 = P6.readsnumber, name1 = "P5", name2 = "P6",
  dummyreads = 5)
}
\arguments{
\item{readsnumber.1}{The first readsnumber input file}

\item{readsnumber.2}{The Second readsnumber input file}

\item{name1}{Give a short name for data 1, such as "P5", "INShi", what ever distinguishable}

\item{name2}{Give a short name for data 2, such as "P5", "INShi", what ever distinguishable}

\item{dummyreads}{Number of dummy reads to add on both in order to alleviate Low reads bias. Derfault is 5}
}
\value{
return a formated merged data.frame with information from both experiment
}
\description{
This function allows you to prepare the data format for downstreme computation
}
\examples{
PrepareData(readsnumber.1=P5.readsnumber,readsnumber.2=P6.readsnumber,name1="P5",name2="P6",dummyreads=5)
}
