% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRNAdistribute.R
\name{sgRNAdistribute}
\alias{sgRNAdistribute}
\title{sgRNAdistribute}
\usage{
sgRNAdistribute(enrichment.data = P5.P6.enrichment)
}
\arguments{
\item{enrichment.data}{The data prepared by "ComputeEnrichment"}
}
\value{
No return but print out useful information of p value and fdr distribution
}
\description{
This function allows you to how the sgRNA distributed in terms of pvalue/fdr and fold change
}
\examples{
sgRNAdistribute(enrichment.data=P5.P6.enrichment)
}
