% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowsgRNAvocan.R
\name{ShowsgRNAvocano}
\alias{ShowsgRNAvocano}
\title{ShowsgRNAvocano}
\usage{
ShowsgRNAvocano(enrichment.data, pCut = 1e-05, fcCut = 2,
  pvalueadjusted = T, pdfname = "sgRNA.vocano.pdf", png = F,
  pngname = "sgRNA.vocano.png")
}
\arguments{
\item{enrichment.data}{The data prepared by "ComputeEnrichment"}

\item{pCut}{p value/fdr cutoff for labeling sgRNAs}

\item{fcCut}{Foldchange cutoff for labeling sgRNAs}

\item{pvalueadjusted}{To tell if the y axis used is the adjusted pvalue or original p value, default is TRUE}

\item{pdfname}{the file name for pdf of vocano plot}

\item{png, }{if true then also make a png format. default is F}

\item{pngname}{the file name for png of vocano plot}
}
\value{
no return but generate a pdf or png
}
\description{
This function plot vocano for sgRNA and label the top sgRNAs
}
\examples{
ShowsgRNAvocano(enrichment.data=P5.P6.enrichment,pCut=1e-5,fcCut=2,pdfname="sgRNA.vocano.pdf",png=T,pngname="sgRNA.vocano.png")
}
