% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_IDs.R
\name{split_IDs}
\alias{split_IDs}
\title{Split the gene ID column into multiple columns using delimiter
specified by split}
\usage{
split_IDs(
  all_gRNAs,
  split = ":",
  column_to_split = "gene",
  new_column_names = c("Symbol", "ID", "gRNA")
)
}
\arguments{
\item{all_gRNAs}{Results from merge_gRNAs}

\item{split}{Used to extract gene symbol, gene_id, and gRNA from c
olumn_to_split of gRNA_count_files. Set split to ":" for extracting gene symbol,
gene ID, and gRNA from : separated string, e.g.,
TMEM192:NM_001100389.1:TGGGTCGTCTTCAATACTCT.}

\item{column_to_split}{The name of the column that contains the
annotation. Default to "gene".}

\item{new_column_names}{Default to c("Symbol", "ID", "gRNA") if
the annotation in column_to_split is in the form of Symbol:ID:gRNA such as
MEM192:NM_001100389.1:TGGGTCGTCTTCAATACTCT, where split is ":"}
}
\description{
Split the gene ID column into multiple columns using delimiter
specified by split
}
\author{
Lihua Julie Zhu
}
