% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create Customized Volcano Plot}
\usage{
volcano_plot(
  x,
  adj_pvalue_col,
  odds_ratio_col,
  adj_pvalue_cutoff = 0.05,
  odds_ratio_cutoff = 1.5,
  genes_to_label,
  label_color = "purple",
  genes_to_highlight,
  highlight_color = "purple",
  gene_col = 1,
  inverse_odds_ratio,
  alpha = 0.5,
  max.overlaps = 80,
  colors = c("purple", "blue", "gray", "orange"),
  legend_color_labels = c("4 gRNAs Altered", "Depleted", "Enriched", "No Change"),
  legend_size_labels = c("0 gRNA Altered", "1 gRNA Altered", paste(2:4, "gRNAs Altered")),
  plot_unit = c("gene", "gRNA"),
  point_size = c(1, 2, 4, 6, 9),
  summarize_FDR_fun = "min",
  summarize_log_odds_ratio_fun = "max",
  highlight_genes_4gRNAs = TRUE,
  label_size = 4
)
}
\arguments{
\item{x}{a data frame such as the output from analyze_CRISPR_screen_data
with minimum of three columns: Gene Identifier, gene symbol,
odds_ratio, and adj_pvalue/FDR}

\item{adj_pvalue_col}{the column where adj_pvalues or FDRs are stored}

\item{odds_ratio_col}{the column where odds_ratios are stored}

\item{adj_pvalue_cutoff}{default to 0.05 which is used for drawing horizontal
line and specifying enriched/depleted gRNAs/genes}

\item{odds_ratio_cutoff}{default to 2 which used for drawing vertical lines
and specifying enriched/depleted gRNAs/genes}

\item{genes_to_label}{the genes that will be labeled in the figure}

\item{label_color}{the color used for highlighting the gene label specified by
genes_to_label}

\item{genes_to_highlight}{the list of genes that will be highlighted in
the figure}

\item{highlight_color}{the color used for highlighting the genes specified
by genes_to_highlight}

\item{gene_col}{the column containing the gene symbol}

\item{inverse_odds_ratio}{indicates whether to reverse the odds_ratio}

\item{alpha}{for specifying the percentage of data to plot}

\item{max.overlaps}{for setting max.overlaps in the function geom_text_repel}

\item{colors}{for specifying colors for depleted, enriched, and unchanged
gRNAs/genes respectively}

\item{legend_color_labels}{for labeling the color legend corresponding to
the colors specified in colors}

\item{legend_size_labels}{for labeling the size legend corresponding to the
size specified in point_size}

\item{plot_unit}{Indicates whether plot the data at gRNA or gene level
gRNA level means plotting one point per gRNA while gene level means plotting
one point per gene with FDR and log_odds_ratio summarized across gRNAs using
functions summarize_FDR_fun and summarize_log_odds_ratio_fun}

\item{point_size}{only applicable when plot_unit is set to gene. Default
setting of c(0.2, 1, 2, 3, 4) to plot point size 0.2, 1, 2, 3, and 4 for
genes with 0, 1, 2, 3, and 4 gRNAs significantly increased/decreased.
Please note that any gene with gRNAs altered significantly in both
directions are removed}

\item{summarize_FDR_fun}{only applicable when plot_unit is set to gene
default to min indicating the minimum FDR of enriched gRNAs for each gene
is used for plotting the enriched genes and the minimum FDR of depleted
gRNAs for each gene is used for plotting depleted ones.}

\item{summarize_log_odds_ratio_fun}{only applicable when plot_unit is set
to gene, default to max indicating that the maximum odds_ratio of each gene
will be used for plotting enriched gene and the minimum odds_ratio will be
used for plotting depleted gene}

\item{highlight_genes_4gRNAs}{only applicable if plot_unit is gene.
It is used for indicating whether to use different color for the genes
with 4 gRNAs signficantly altered in the same direction.}

\item{label_size}{size of the genes to highlight in the plot}
}
\value{
a ggplot object
}
\description{
Create Customized Volcano Plot
}
\examples{

x <- readRDS(system.file("extdata", "example_data.RDS",
            package = "CRISPRscreen"))
genes_to_label = c("GPAA1")

# Plot one point for each gene by setting plot_unit to "gene"
volcano_plot(x, adj_pvalue_col=12, odds_ratio_col=11,
       odds_ratio_cutoff = 1.5,
       genes_to_label = genes_to_label,
       genes_to_highlight = genes_to_label,  plot_unit = "gene",
       gene_col = 2, highlight_color = "red", label_color ="blue",
      inverse_odds_ratio = FALSE)

#  Plot one point for each gRNA by setting plot_unit to "gRNA"
volcano_plot(x, adj_pvalue_col=12, odds_ratio_col=11,
       genes_to_label = genes_to_label,
       genes_to_highlight = genes_to_label,  plot_unit = "gRNA",
       gene_col = 2, highlight_color = "red", label_color ="blue",
      inverse_odds_ratio = FALSE)

}
\author{
Lihua Julie Zhu
}
