% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iplot}
\alias{iplot}
\title{Interactive scatter plot using plotly (simple wrapper)}
\usage{
iplot(x, y, labels, xlab = "", ylab = "", title = "", col = NULL, pal = NULL)
}
\arguments{
\item{x}{Numeric vector of x values}

\item{y}{Numeric vector of y values}

\item{labels}{Data labels}

\item{xlab}{String for x-axis label}

\item{ylab}{String for y-axis label}

\item{title}{String for plot title}

\item{col}{Numeric vector of length of 1, or equal to the length of x, specifying points color}

\item{pal}{Character vector specifying color names}
}
\value{
An interactive plot as a side value.
}
\description{
Interactive scatter plot using plotly (simple wrapper).
    Used mainly by JP for quick plots.
}
\author{
Jean-Philippe Fortin
}
